fix(consultation): fixed data type.

+++ src/main/java/com/loremipsum/lawconnectplatform/consultation/domain/model/valueobjects/ClientC.java
@@ -3,5 +3,5 @@ package com.loremipsum.lawconnectplatform.consultation.domain.model.valueobjects
 import jakarta.persistence.Embeddable;
 
 @Embeddable
-public record ClientC(int clientId) {
+public record ClientC(Long clientId) {
 }
