+++ src/main/java/com/loremipsum/lawconnectplatform/consultation/domain/model/valueobjects/LawyerC.java
@@ -3,5 +3,5 @@ package com.loremipsum.lawconnectplatform.consultation.domain.model.valueobjects
 import jakarta.persistence.Embeddable;
 
 @Embeddable
-public record LawyerC(int lawyerId) {
+public record LawyerC(Long lawyerId) {
 }
